Crate datafusion_optimizer

Re-exports

pub use optimizer::OptimizerConfig;

pub use optimizer::OptimizerContext;

pub use optimizer::OptimizerRule;

pub use utils::optimize_children;

Modules

alias

common_subexpr_eliminate

Eliminate common sub-expression.

decorrelate_where_exists

decorrelate_where_in

eliminate_cross_join

Optimizer rule to eliminate cross join to inner join if join predicates are available in filters.

eliminate_filter

Optimizer rule to replace where false on a plan with an empty relation. This saves time in planning and executing the query. Note that this rule should be applied after simplify expressions optimizer rule.

eliminate_limit

Optimizer rule to replace LIMIT 0 or LIMIT whose ancestor LIMIT's skip is greater than or equal to current's fetch on a plan with an empty relation. This rule also removes OFFSET 0 from the LogicalPlan This saves time in planning and executing the query.

eliminate_outer_join

Optimizer rule to eliminate left/right/full join to inner join if possible.

extract_equijoin_predicate

Optimizer rule to extract equijoin expr from filter

filter_null_join_keys

The FilterNullJoinKeys rule will identify inner joins with equi-join conditions where the join key is nullable on one side and non-nullable on the other side and then insert an IsNotNull filter on the nullable side since null values can never match.

inline_table_scan

Optimizer rule to replace TableScan references such as DataFrames and Views and inlines the LogicalPlan to support further optimization

optimizer

Query optimizer traits

propagate_empty_relation

push_down_filter

Push Down Filter optimizer rule ensures that filters are applied as early as possible in the plan

push_down_limit

Optimizer rule to push down LIMIT in the query plan It will push down through projection, limits (taking the smaller limit)

push_down_projection

Projection Push Down optimizer rule ensures that only referenced columns are loaded into memory

rewrite_disjunctive_predicate

rewrite_predicate_cnf

scalar_subquery_to_join

simplify_expressions

single_distinct_to_groupby

single distinct to group by optimizer rule

type_coercion

Optimizer rule for type validation and coercion

unwrap_cast_in_comparison

Unwrap-cast binary comparison rule can be used to the binary/inlist comparison expr now, and other type of expr can be added if needed. This rule can reduce adding the Expr::Cast the expr instead of adding the Expr::Cast to literal expr.

utils

Collection of utility functions that are leveraged by the query optimizer rules