Trait datafusion_optimizer::optimizer::OptimizerRule

pub trait OptimizerRule {
    // Required methods
    fn try_optimize(
        &self,
        plan: &LogicalPlan,
        config: &dyn OptimizerConfig
    ) -> Result<Option<LogicalPlan>>;
    fn name(&self) -> &str;

    // Provided method
    fn apply_order(&self) -> Option<ApplyOrder> { ... }
}

OptimizerRule transforms one [‘LogicalPlan’] into another which computes the same results, but in a potentially more efficient way. If there are no suitable transformations for the input plan, the optimizer can simply return it as is.

Required Methods

fn try_optimize( &self, plan: &LogicalPlan, config: &dyn OptimizerConfig ) -> Result<Option<LogicalPlan>>

Try and rewrite plan to an optimized form, returning None if the plan cannot be optimized by this rule.

fn name(&self) -> &str

A human readable name for this optimizer rule

Provided Methods

fn apply_order(&self) -> Option<ApplyOrder>

How should the rule be applied by the optimizer? See comments on ApplyOrder for details.

If a rule use default None, its should traverse recursively plan inside itself