Trait datafusion_optimizer::simplify_expressions::context::SimplifyInfo

pub trait SimplifyInfo {
    // Required methods
    fn is_boolean_type(&self, expr: &Expr) -> Result<bool>;
    fn nullable(&self, expr: &Expr) -> Result<bool>;
    fn execution_props(&self) -> &ExecutionProps;
}

The information necessary to apply algebraic simplification to an Expr. See SimplifyContext for one concrete implementation.

This trait exists so that other systems can plug schema information in without having to create DFSchema objects. If you have a DFSchemaRef you can use SimplifyContext

Required Methods

fn is_boolean_type(&self, expr: &Expr) -> Result<bool>

returns true if this Expr has boolean type

fn nullable(&self, expr: &Expr) -> Result<bool>

returns true of this expr is nullable (could possibly be NULL)

fn execution_props(&self) -> &ExecutionProps

Returns details needed for partial expression evaluation