Module datafusion_optimizer::utils

Collection of utility functions that are leveraged by the query optimizer rules

Functions

add_filter

returns a new LogicalPlan that wraps plan in a LogicalPlan::Filter with its predicate be all predicates ANDed.

conjunction

Combines an array of filter expressions into a single filter expression consisting of the input filter expressions joined with logical AND.

disjunction

Combines an array of filter expressions into a single filter expression consisting of the input filter expressions joined with logical OR.

exprs_to_join_cols

Extracts correlating columns from expressions

find_join_exprs

Looks for correlating expressions: equality expressions with one field from the subquery, and one not in the subquery (closed upon from outer scope)

merge_schema

merge inputs schema into a single schema.

only_or_err

Returns the first (and only) element in a slice, or an error

optimize_children

Convenience rule for writing optimizers: recursively invoke optimize on plan’s children and then return a node of the same type. Useful for optimizer rules which want to leave the type of plan unchanged but still apply to the children. This also handles the case when the plan is a LogicalPlan::Explain.

rewrite_preserving_name

Rewrites expr using rewriter, ensuring that the output has the same name as expr prior to rewrite, adding an alias if necessary.

split_binary

Splits an binary operator tree Expr such as A <OP> B <OP> C => [A, B, C]

split_binary_owned

Splits an owned binary operator tree Expr such as A <OP> B <OP> C => [A, B, C]

split_conjunction

Splits a conjunctive Expr such as A AND B AND C => [A, B, C]

split_conjunction_owned

Splits an owned conjunctive Expr such as A AND B AND C => [A, B, C]

unalias

Recursively un-alias an expressions

verify_not_disjunction

Recursively scans a slice of expressions for any Or operators